services/recurring-task/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
python-dateutil>=2.9.0
